Melhorando tratamento de erros com http-error

+++ src/services/UsuarioService.ts
@@ -1,5 +1,7 @@
-import { UsuarioError } from '../errors/UsuarioError';
+import { validate } from 'uuid';
 import { UsuarioRepository } from '../repository/usuario';
+import { NotFound, BadRequest } from 'http-errors';
+import { NextFunction } from 'express';
 
 interface IUser {
 	nome: string;
@@ -10,40 +12,48 @@ interface IUser {
 }
 
 export class UsuarioService {
-	async createUser(usuario: IUser) {
+	async createUser(usuario: IUser, next: NextFunction) {
 		const alreadyExists = await UsuarioRepository.findOneBy({
 			nome: usuario.nome,
 			sobrenome: usuario.sobrenome,
 		});
 
-		if (alreadyExists) return UsuarioError.userAlreadyExists();
+		if (alreadyExists) return next(new BadRequest('Usuario ja cadastrado.'));
 
 		for (const [key, value] of Object.entries(usuario)) {
-			if (value.length === 0) return UsuarioError.emptyField(key);
+			const field = key.toUpperCase();
+			if (value.length === 0)
+				return next(
+					new BadRequest(`O campo ${field} nao pode estar vazio.`)
+				);
 		}
 
-		const createdUser = await UsuarioRepository.save(usuario);
-		return createdUser;
+		return UsuarioRepository.save(usuario);
 	}
 
 	async getUsers() {
-		const users = await UsuarioRepository.find();
-		return users;
+		return await UsuarioRepository.find();
 	}
 
-	async getUserById(id: string) {
+	async getUserById(id: string, next: NextFunction) {
+		if (!validate(id)) return next(new NotFound('ID invalido'));
+
 		const user = await UsuarioRepository.findOneBy({ id: id });
 
-		if (!user) return UsuarioError.userNotFound();
+		if (!user) return next(new NotFound('Usuario nao encontrado.'));
 
 		return user;
 	}
 
-	async updateUser(id: string, data: {}) {
+	async updateUser(id: string, data: {}, next: NextFunction) {
+		if (!validate(id)) return next(new NotFound('ID invalido'));
+
 		return await UsuarioRepository.update(id, data);
 	}
 
-	async deleteUser(id: string) {
+	async deleteUser(id: string, next: NextFunction) {
+		if (!validate(id)) return next(new NotFound('ID invalido'));
+
 		return await UsuarioRepository.delete(id);
 	}
 }
